fix pr notes
